+++ optmod/coptmod/node.h
@@ -14,6 +14,7 @@
 #include <stdlib.h>
 #include <math.h>
 #include "uthash.h"
+#include <inttypes.h>
 
 #define NODE_BUFFER_SIZE 100
 
@@ -34,16 +35,16 @@ Node* NODE_array_get(Node* n, int i);
 void NODE_array_del(Node* n, int num);
 void NODE_copy_from_node(Node* n, Node* other, Node* hash);
 int NODE_get_index(Node* n);
-long NODE_get_id(Node* n);
+uintptr_t NODE_get_id(Node* n);
 int NODE_get_type(Node* n);
 double NODE_get_value(Node* n);
 char* NODE_get_type_name(Node* n);
 Node* NODE_hash_add(Node* hash, Node* n);
-Node* NODE_hash_find(Node* hash, long id);
+Node* NODE_hash_find(Node* hash, uintptr_t id);
 void NODE_hash_del(Node* hash);
 void NODE_init(Node* n);
 void NODE_set_type(Node* n, int type);
-void NODE_set_id(Node* n, long id);
+void NODE_set_id(Node* n, uintptr_t id);
 void NODE_set_value(Node* n, double value);
 void NODE_set_arg1(Node* n, Node* arg1);
 void NODE_set_arg2(Node* n, Node* arg2);
